feat: agent support for podman
